backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
ollama
chromadb
python-docx
pdfplumber
sentence-transformers
python-multipart
